@@ -31,13 +31,13 @@
         }
     ],
     "require": {
-        "php": "^8.1",
-        "illuminate/support": "^10.0"
+        "php": "^8.2",
+        "illuminate/support": "^11.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.1",
-        "orchestra/testbench": "^8.5",
-        "phpunit/phpunit": "^10.0",
+        "orchestra/testbench": "^9.0",
+        "phpunit/phpunit": "^11.1",
         "squizlabs/php_codesniffer": "^3.5.2"
     },
     "autoload": {
